FA sniffer: Fix bug in double decimation calculation
Could lead to calculation of negative square root and thus a crash.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/drivers/pcie_sniffer@50977 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tools/transform.c
@@ -249,8 +249,8 @@ static void decimate_decimation(
     {
         sumx += input->mean.x;
         sumy += input->mean.y;
-        sumvarx = input->std.x * input->std.x;
-        sumvary = input->std.y * input->std.y;
+        sumvarx = (double) input->std.x * input->std.x;
+        sumvary = (double) input->std.y * input->std.y;
         if (input->min.x < minx)     minx = input->min.x;
         if (maxx < input->max.x)     maxx = input->max.x;
         if (input->min.y < miny)     miny = input->min.y;
